refactor: move GroupThemeToLabel to themeUtils
