+++ src/components/ThemeBuilder/ThemeBuilder.tsx
@@ -2,7 +2,8 @@ import { HTMLAttributes } from "react";
 import { HexColorPicker } from "react-colorful";
 
 import { makeClassName } from "@/utils/tsxUtils";
-import { GroupTheme, GroupThemeToLabel } from "@/types/Group";
+import { GroupThemeToLabel } from "@/utils/themeUtils";
+import { GroupTheme } from "@/types/Group";
 
 import styles from "./ThemeBuilder.module.less";
