+++ src/utils/themeUtils.ts
@@ -25,3 +25,21 @@ export const GroupThemeToCssVariable: Record<keyof GroupTheme, string> = {
     fontColorSuccess: "--font-color-success",
     wrongInvalidColor: "--wrong-invalid-color",
 };
+
+export const GroupThemeToLabel: Record<keyof GroupTheme, string> = {
+    // TODO: Name all variables
+    primaryColor:           "Cor primária",
+    secondaryColor:         "Cor secundária",
+    backgroundColor:        "Cor de fundo",
+    cardBackgroundColor:    "Cor de fundo de cartões",
+    cardItemColor:          "Cor de de itens de cartões",
+    fontColorV1:            "Cor de fonte 1",
+    fontColorV2:            "Cor de fonte 2",
+    fontColorV3:            "Cor de fonte 3",
+    fontColorLinks:         "Cor de fonte de links",
+    fontColorDisabled:      "Cor de fonte desabilitada",
+    buttonHoverColor:       "buttonHoverColor",
+    fontColorAlert:         "Cor de alerta",
+    fontColorSuccess:       "Cor de sucesso",
+    wrongInvalidColor:      "Cor de erro",
+};
